fix buggy logic in Select: now click outside works
